Backend/requirements.txt
pandas==1.0.3
Flask_SQLAlchemy==2.4.1
Flask==1.1.2
tensorflow==2.0.0
Flask_APScheduler==1.11.0
matplotlib==3.1.1
numpy==1.17.0
requests==2.22.0
Flask_Cors==3.0.8
beautifulsoup4==4.9.0
python-decouple==3.3
